move the language logic from the Project to the AnalysisInputLocation
