+++ sootup.examples/src/test/java/sootup/examples/basicSetup/BasicSetup.java
@@ -67,18 +67,17 @@ public class BasicSetup {
     Project project =
         JavaProject.builder((JavaLanguage) language).addInputLocation(inputLocation).build();
 
+    // Create a view for project, which allows us to retrieve classes
+    View view = project.createView();
+
     // Create a signature for the class we want to analyze
-    ClassType classType = project.getIdentifierFactory().getClassType("HelloWorld");
+    ClassType classType = view.getIdentifierFactory().getClassType("HelloWorld");
 
     // Create a signature for the method we want to analyze
     MethodSignature methodSignature =
-        project
-            .getIdentifierFactory()
+        view.getIdentifierFactory()
             .getMethodSignature(
                 classType, "main", "void", Collections.singletonList("java.lang.String[]"));
-
-    // Create a view for project, which allows us to retrieve classes
-    View view = project.createView();
 
     // Assert that class is present
     assertTrue(view.getClass(classType).isPresent());
