+++ sootup.java.bytecode/src/main/java/sootup/java/bytecode/inputlocation/PathBasedAnalysisInputLocation.java
@@ -1,15 +1,8 @@
 package sootup.java.bytecode.inputlocation;
 
-import com.google.common.cache.CacheBuilder;
-import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
-import com.google.common.cache.RemovalNotification;
-import com.googlecode.dex2jar.tools.Dex2jarCmd;
 import java.io.*;
 import java.nio.file.*;
 import java.util.*;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
 import java.util.jar.Attributes;
 import java.util.jar.JarInputStream;
 import java.util.jar.Manifest;
@@ -38,11 +31,8 @@ import sootup.core.util.PathUtils;
 import sootup.core.util.StreamUtils;
 import sootup.core.views.View;
 import sootup.java.bytecode.frontend.AsmJavaClassProvider;
-import sootup.java.bytecode.frontend.AsmModuleSource;
 import sootup.java.core.*;
-import sootup.java.core.signatures.ModuleSignature;
 import sootup.java.core.types.JavaClassType;
-import sootup.java.core.types.ModuleJavaClassType;
 
 /*-
  * #%L
@@ -98,10 +88,6 @@ public abstract class PathBasedAnalysisInputLocation
     } else if (PathUtils.isArchive(path)) {
       if (PathUtils.hasExtension(path, FileType.WAR)) {
         inputLocation = new WarArchiveAnalysisInputLocation(path, srcType);
-      } else if (isMultiReleaseJar(path)) { // check if mainfest contains multi release flag
-        inputLocation = new MultiReleaseJarAnalysisInputLocation(path, srcType);
-      } else if (PathUtils.hasExtension(path, FileType.APK)) {
-        inputLocation = new ApkAnalysisInputLocation(path, srcType);
       } else {
         inputLocation = new ArchiveBasedAnalysisInputLocation(path, srcType);
       }
@@ -244,361 +230,6 @@ public abstract class PathBasedAnalysisInputLocation
     public Optional<? extends AbstractClassSource<JavaSootClass>> getClassSource(
         @Nonnull ClassType type, @Nonnull View<?> view) {
       return getClassSourceInternal((JavaClassType) type, path, new AsmJavaClassProvider(view));
-    }
-  }
-
-  public static class MultiReleaseJarAnalysisInputLocation extends ArchiveBasedAnalysisInputLocation
-      implements ModuleInfoAnalysisInputLocation {
-
-    @Nonnull private final int[] availableVersions;
-
-    @Nonnull
-    private final Map<Integer, Map<ModuleSignature, JavaModuleInfo>> moduleInfoMap =
-        new HashMap<>();
-
-    @Nonnull
-    private final Map<Integer, List<AnalysisInputLocation<JavaSootClass>>> inputLocations =
-        new HashMap<>();
-
-    @Nonnull
-    private final List<AnalysisInputLocation<JavaSootClass>> baseInputLocations = new ArrayList<>();
-
-    boolean isResolved = false;
-
-    private MultiReleaseJarAnalysisInputLocation(@Nonnull Path path, @Nullable SourceType srcType) {
-      super(path, srcType);
-
-      int[] tmp;
-      try {
-        FileSystem fs = fileSystemCache.get(path);
-        final Path archiveRoot = fs.getPath("/");
-        tmp =
-            Files.list(archiveRoot.getFileSystem().getPath("/META-INF/versions/"))
-                .map(dir -> dir.getFileName().toString().replace("/", ""))
-                .mapToInt(Integer::new)
-                .sorted()
-                .toArray();
-      } catch (IOException | ExecutionException e) {
-        e.printStackTrace();
-        tmp = new int[] {};
-      }
-      availableVersions = tmp;
-
-      discoverInputLocations(srcType);
-    }
-
-    /** Discovers all input locations for different java versions in this multi release jar */
-    private void discoverInputLocations(@Nullable SourceType srcType) {
-      FileSystem fs = null;
-      try {
-        fs = fileSystemCache.get(path);
-      } catch (ExecutionException e) {
-        e.printStackTrace();
-      }
-      final Path archiveRoot = fs.getPath("/");
-      final String moduleInfoFilename = JavaModuleIdentifierFactory.MODULE_INFO_FILE + ".class";
-
-      baseInputLocations.add(PathBasedAnalysisInputLocation.create(archiveRoot, srcType));
-
-      String sep = archiveRoot.getFileSystem().getSeparator();
-
-      if (!isResolved) {
-
-        for (int i = availableVersions.length - 1; i >= 0; i--) {
-          inputLocations.put(availableVersions[i], new ArrayList<>());
-
-          final Path versionRoot =
-              archiveRoot
-                  .getFileSystem()
-                  .getPath("/META-INF/versions/" + availableVersions[i] + sep);
-
-          // only versions >= 9 support java modules
-          if (availableVersions[i] > 8) {
-            moduleInfoMap.put(availableVersions[i], new HashMap<>());
-            try (DirectoryStream<Path> stream = Files.newDirectoryStream(versionRoot)) {
-              for (Path entry : stream) {
-
-                Path mi = path.resolve(moduleInfoFilename);
-
-                if (Files.exists(mi)) {
-                  JavaModuleInfo moduleInfo = new AsmModuleSource(mi);
-                  ModuleSignature moduleSignature = moduleInfo.getModuleSignature();
-                  JavaModulePathAnalysisInputLocation inputLocation =
-                      new JavaModulePathAnalysisInputLocation(
-                          versionRoot.toString(), versionRoot.getFileSystem(), getSourceType());
-
-                  inputLocations.get(availableVersions[i]).add(inputLocation);
-                  moduleInfoMap.get(availableVersions[i]).put(moduleSignature, moduleInfo);
-                }
-
-                if (Files.isDirectory(entry)) {
-                  mi = versionRoot.resolve(moduleInfoFilename);
-
-                  if (Files.exists(mi)) {
-                    JavaModuleInfo moduleInfo = new AsmModuleSource(mi);
-                    ModuleSignature moduleSignature = moduleInfo.getModuleSignature();
-                    JavaModulePathAnalysisInputLocation inputLocation =
-                        new JavaModulePathAnalysisInputLocation(
-                            versionRoot.toString(), versionRoot.getFileSystem(), getSourceType());
-
-                    inputLocations.get(availableVersions[i]).add(inputLocation);
-                    moduleInfoMap.get(availableVersions[i]).put(moduleSignature, moduleInfo);
-                  }
-                  // else TODO [bh] can we have automatic modules here?
-                }
-              }
-            } catch (IOException e) {
-              e.printStackTrace();
-            }
-          }
-
-          // if there was no module or the version is not > 8, we just add a directory based input
-          // location
-          if (inputLocations.get(availableVersions[i]).size() == 0) {
-            inputLocations
-                .get(availableVersions[i])
-                .add(PathBasedAnalysisInputLocation.create(versionRoot, srcType));
-          }
-        }
-      }
-
-      isResolved = true;
-    }
-
-    @Override
-    @Nonnull
-    public Optional<? extends AbstractClassSource<JavaSootClass>> getClassSource(
-        @Nonnull ClassType type, @Nonnull View<?> view) {
-
-      Collection<AnalysisInputLocation<JavaSootClass>> il =
-          getBestMatchingInputLocationsRaw(view.getProject().getLanguage().getVersion());
-
-      Collection<AnalysisInputLocation<JavaSootClass>> baseIl = getBaseInputLocations();
-
-      if (type instanceof ModuleJavaClassType) {
-        il =
-            il.stream()
-                .filter(location -> location instanceof ModuleInfoAnalysisInputLocation)
-                .collect(Collectors.toList());
-        baseIl =
-            baseIl.stream()
-                .filter(location -> location instanceof ModuleInfoAnalysisInputLocation)
-                .collect(Collectors.toList());
-      } else {
-        il =
-            il.stream()
-                .filter(location -> !(location instanceof ModuleInfoAnalysisInputLocation))
-                .collect(Collectors.toList());
-        baseIl =
-            baseIl.stream()
-                .filter(location -> !(location instanceof ModuleInfoAnalysisInputLocation))
-                .collect(Collectors.toList());
-      }
-
-      Optional<? extends AbstractClassSource<JavaSootClass>> foundClass =
-          il.stream()
-              .map(location -> location.getClassSource(type, view))
-              .filter(Optional::isPresent)
-              .limit(1)
-              .map(Optional::get)
-              .findAny();
-
-      if (foundClass.isPresent()) {
-        return foundClass;
-      } else {
-        return baseIl.stream()
-            .map(location -> location.getClassSource(type, view))
-            .filter(Optional::isPresent)
-            .limit(1)
-            .map(Optional::get)
-            .findAny();
-      }
-    }
-
-    @Nonnull
-    @Override
-    public Collection<? extends AbstractClassSource<JavaSootClass>> getModulesClassSources(
-        @Nonnull ModuleSignature moduleSignature, @Nonnull View<?> view) {
-      return inputLocations.get(view.getProject().getLanguage().getVersion()).stream()
-          .filter(location -> location instanceof ModuleInfoAnalysisInputLocation)
-          .map(
-              location ->
-                  ((ModuleInfoAnalysisInputLocation) location)
-                      .getModulesClassSources(moduleSignature, view))
-          .flatMap(Collection::stream)
-          .collect(Collectors.toList());
-    }
-
-    /**
-     * Returns the best matching input locations or the base input location.
-     *
-     * @param javaVersion version to find best match to
-     * @return best match or base input locations
-     */
-    private Collection<AnalysisInputLocation<JavaSootClass>> getBestMatchingInputLocationsRaw(
-        int javaVersion) {
-      for (int i = availableVersions.length - 1; i >= 0; i--) {
-
-        if (availableVersions[i] > javaVersion) continue;
-
-        return new ArrayList<>(inputLocations.get(availableVersions[i]));
-      }
-
-      return getBaseInputLocations();
-    }
-
-    private Collection<AnalysisInputLocation<JavaSootClass>> getBaseInputLocations() {
-      return baseInputLocations;
-    }
-
-    @Override
-    @Nonnull
-    public Collection<? extends AbstractClassSource<JavaSootClass>> getClassSources(
-        @Nonnull View<?> view) {
-      Collection<AnalysisInputLocation<JavaSootClass>> il =
-          getBestMatchingInputLocationsRaw(view.getProject().getLanguage().getVersion());
-
-      Collection<AbstractClassSource<JavaSootClass>> result =
-          il.stream()
-              .map(location -> location.getClassSources(view))
-              .flatMap(Collection::stream)
-              .collect(Collectors.toList());
-
-      if (il != getBaseInputLocations()) {
-
-        Collection<AbstractClassSource<JavaSootClass>> baseSources =
-            getBaseInputLocations().stream()
-                .map(location -> location.getClassSources(view))
-                .flatMap(Collection::stream)
-                .collect(Collectors.toList());
-
-        baseSources.forEach(
-            cs -> {
-              // do not add duplicate class sources
-              if (result.stream()
-                  .noneMatch(
-                      bestMatchCS ->
-                          bestMatchCS
-                              .getClassType()
-                              .getFullyQualifiedName()
-                              .equals(cs.getClassType().getFullyQualifiedName()))) {
-                result.add(cs);
-              }
-            });
-      }
-
-      return result;
-    }
-
-    @Nonnull
-    @Override
-    public Optional<JavaModuleInfo> getModuleInfo(ModuleSignature sig, View<?> view) {
-      return Optional.ofNullable(
-          moduleInfoMap.get(view.getProject().getLanguage().getVersion()).get(sig));
-    }
-
-    @Nonnull
-    @Override
-    public Set<ModuleSignature> getModules(View<?> view) {
-      return inputLocations.get(view.getProject().getLanguage().getVersion()).stream()
-          .filter(e -> e instanceof ModuleInfoAnalysisInputLocation)
-          .map(e -> ((ModuleInfoAnalysisInputLocation) e).getModules(view))
-          .flatMap(Set::stream)
-          .collect(Collectors.toSet());
-    }
-
-    @Override
-    public boolean equals(Object o) {
-      if (!(o instanceof PathBasedAnalysisInputLocation)) {
-        return false;
-      }
-      return path.equals(((PathBasedAnalysisInputLocation) o).path);
-    }
-
-    @Override
-    public int hashCode() {
-      return path.hashCode();
-    }
-  }
-
-  private static class ApkAnalysisInputLocation extends ArchiveBasedAnalysisInputLocation {
-
-    private ApkAnalysisInputLocation(@Nonnull Path path, @Nullable SourceType srcType) {
-      super(path, srcType);
-      String jarPath = dex2jar(path);
-      this.path = Paths.get(jarPath);
-    }
-
-    private String dex2jar(Path path) {
-      String apkPath = path.toAbsolutePath().toString();
-      String outDir = "./tmp/";
-      int start = apkPath.lastIndexOf(File.separator);
-      int end = apkPath.lastIndexOf(".apk");
-      String outputFile = outDir + apkPath.substring(start + 1, end) + ".jar";
-      Dex2jarCmd.main("-f", apkPath, "-o", outputFile);
-      return outputFile;
-    }
-  }
-
-  private static class ArchiveBasedAnalysisInputLocation extends PathBasedAnalysisInputLocation {
-
-    // We cache the FileSystem instances as their creation is expensive.
-    // The Guava Cache is thread-safe (see JavaDoc of LoadingCache) hence this
-    // cache can be safely shared in a static variable.
-    protected static final LoadingCache<Path, FileSystem> fileSystemCache =
-        CacheBuilder.newBuilder()
-            .removalListener(
-                (RemovalNotification<Path, FileSystem> removalNotification) -> {
-                  try {
-                    removalNotification.getValue().close();
-                  } catch (IOException e) {
-                    throw new RuntimeException(
-                        "Could not close file system of " + removalNotification.getKey(), e);
-                  }
-                })
-            .expireAfterAccess(1, TimeUnit.SECONDS)
-            .build(
-                CacheLoader.from(
-                    path -> {
-                      try {
-                        return FileSystems.newFileSystem(
-                            Objects.requireNonNull(path), (ClassLoader) null);
-                      } catch (IOException e) {
-                        throw new RuntimeException("Could not open file system of " + path, e);
-                      }
-                    }));
-
-    private ArchiveBasedAnalysisInputLocation(@Nonnull Path path, @Nullable SourceType srcType) {
-      super(path, srcType);
-    }
-
-    @Override
-    @Nonnull
-    public Optional<? extends AbstractClassSource<JavaSootClass>> getClassSource(
-        @Nonnull ClassType type, @Nonnull View<?> view) {
-      try {
-        FileSystem fs = fileSystemCache.get(path);
-        final Path archiveRoot = fs.getPath("/");
-        return getClassSourceInternal(
-            (JavaClassType) type, archiveRoot, new AsmJavaClassProvider(view));
-      } catch (ExecutionException e) {
-        throw new RuntimeException("Failed to retrieve file system from cache for " + path, e);
-      }
-    }
-
-    @Override
-    @Nonnull
-    public Collection<? extends AbstractClassSource<JavaSootClass>> getClassSources(
-        @Nonnull View<?> view) {
-      // we don't use the filesystem cache here as it could close the filesystem after the timeout
-      // while we are still iterating
-      try (FileSystem fs = FileSystems.newFileSystem(path, (ClassLoader) null)) {
-        final Path archiveRoot = fs.getPath("/");
-        return walkDirectory(
-            archiveRoot, view.getProject().getIdentifierFactory(), new AsmJavaClassProvider(view));
-      } catch (IOException e) {
-        throw new RuntimeException(e);
-      }
     }
   }
